Use a .env file if present
It's common to put environment variables in a .env file
Using the `dtenv` gem or another way to load these values
makes it easy to use them in the app, including the database configuration.

The use of an Env class is heavily inspired by Foreman
cf. https://github.com/ddollar/foreman/commit/19e9806c4cdf348325fe8312fac827774af4affa

In Foreman's case, the file is referenced by its name.
Here we have to capture the file's content then pass it to the class,
since we're on different locations.